Changed app icon, notified master controller when underlying detail page changes

+++ NoteTaker.iOS/ViewControllers/DetailViewController.cs
@@ -2,7 +2,7 @@
 using NoteTaker.Core.Models;
 using NoteTaker.Core.Services;
 using System;
-
+using NoteTaker.Core.Enums;
 using UIKit;
 
 namespace NoteTaker.iOS
@@ -12,6 +12,8 @@ namespace NoteTaker.iOS
         public INoteStorageService NoteStorageService { get; set; }
 
 		public NoteEntryModel Note { get; set; }
+
+        public Action<NoteEntryModel, NoteOperationType> OnEntityChanged { get; set; }
 
 		public DetailViewController (IntPtr handle) : base (handle)
 		{
@@ -47,16 +49,42 @@ namespace NoteTaker.iOS
             var saveButton = new UIBarButtonItem(UIBarButtonSystemItem.Done, OnSave) { AccessibilityLabel = "saveButton" };
             var deleteButton = new UIBarButtonItem(UIBarButtonSystemItem.Trash, OnDelete) { AccessibilityLabel = "deleteButton" };
             NavigationItem.RightBarButtonItems = new UIBarButtonItem[] { saveButton, deleteButton };
+
+            TitleTextView.AddTarget(TitleTextViewOnValueChanged, UIControlEvent.EditingChanged);// += TitleTextViewOnValueChanged;
+            NoteDescriptionTextView.Changed += NoteDescriptionTextViewOnChanged;
         }
 
-        private async void OnDelete(object sender, EventArgs e)
+	    private void NoteDescriptionTextViewOnChanged(object sender, EventArgs eventArgs)
+	    {
+	        Note.Text = ((UITextView)sender).Text;
+        }
+
+	    private void TitleTextViewOnValueChanged(object sender, EventArgs eventArgs)
+	    {
+
+	        Note.Title = ((UITextField) sender).Text;
+	    }
+
+	    private async void OnDelete(object sender, EventArgs e)
         {
+            if (Note == null)
+            {
+                return;
+            }
+
             var result = await NoteStorageService.RemoveNote(Note);
 
             if (result)
             {
                 UserDialogs.Instance.Toast("Note deleted", TimeSpan.FromSeconds(3));
                 //notify masterviewcontroller
+                OnEntityChanged(Note, NoteOperationType.Delete);
+
+                DismissKeyboard();
+            }
+            else
+            {
+                UserDialogs.Instance.Toast($"An error occurred while deleting note: {Note.Title}. Please try again");
             }
         }
 
@@ -77,13 +105,14 @@ namespace NoteTaker.iOS
                 {
                     UserDialogs.Instance.Toast("Note saved", TimeSpan.FromSeconds(3));
                     //notify masterviewcontroller
+                    OnEntityChanged?.Invoke(Note, NoteOperationType.AddOrEdit);
                 }
+
+                DismissKeyboard();
             }
             else
             {
-                var alert = UIAlertController.Create("Warning", exisingNoteText, UIAlertControllerStyle.Alert);
-                alert.AddAction(UIAlertAction.Create("OK", UIAlertActionStyle.Default, null));
-                PresentViewController(alert, true, null);
+                UserDialogs.Instance.Toast($"Warning {exisingNoteText}.");
             }
 
         }
@@ -93,6 +122,19 @@ namespace NoteTaker.iOS
 			base.DidReceiveMemoryWarning ();
 			// Release any cached data, images, etc that aren't in use.
 		}
+
+	    private void DismissKeyboard()
+	    {
+	        if (TitleTextView.CanResignFirstResponder)
+	        {
+	            TitleTextView.ResignFirstResponder();
+	        }
+
+	        if (NoteDescriptionTextView.CanResignFirstResponder)
+	        {
+	            NoteDescriptionTextView.ResignFirstResponder();
+	        }
+        }
 	}
 }
